This is a slightly revised implementation of removeDeadSymbols that "resurrects" type symbols that are considered dead by analysis, but which are necessary to code generation. A simple example of this would be:
	class myclass { ... }

	var c: myclass;

	if (c == nil) {
	  writeln("c is nil!");
	}

Here, analysis considers myclass dead because
no instantiations of it are ever made.  However,
c is not dead, so it will be codegened to be
a C variable of type myclass.  Yet myclass would
not have been code generated.  So in this case,
c's liveness causes myclass to be resurrected.

Unfortunately, having written this, I think
that my approach was too simple, as in general
we may need to iterate until we reach a fixed
point.  So this is a bit of a stopgap.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@2001 3a8e244f-b0f2-452b-bcba-4c88e055c3ca