This change removes all tabs from code, replacing them with 8 spaces (the same as emacs windows will tend to show by default).
We've decided to disallow the use of tabs in code for
all Chapel source in order to make the code appear
identical in all editors.  New code should not introduce
new tabs.  Code should be indented using spaces, and the
default indentation amount is two spaces per line.

By default, emacs compresses groups of 8 spaces into a
tab when you hit tab in certain editing modes, so most
of us have been inserting tabs all along without knowing
it.

If you are an emacs user, drop the following line into
your .emacs file *now*:

     (setq-default indent-tabs-mode nil)

(John, note that this is slightly different than what
I think we mailed the other day, which only sets it
for the current buffer).

If you have modifications in progress currently, you
will want to remove the tabs from it.  In order to
see whether the code has tabs, do:

    grep -P '\t' <filename>

In order to remove the tabs from the code, you can either:

   load into emacs, select the whole file (Ctrl-space then
   Ctrl->), then do M-x untabify

Or:

   run expand from the unix command line on the file, piping
   the result to a temporary file, then copying the temporary
   file back into place (the following script does this:
-----------------------------------------------
#!/bin/csh

if ($# == 0) then
  echo "Usage: $0 <filename>"
  exit 0
endif

expand $1 > /tmp/tmp.tmp && cp /tmp/tmp.tmp $1
-----------------------------------------------

If you need to do a bunch of files, put the script above
into an exeutable location, name it tabs2spaces, and run
a command like the following from your root chapel directory:

          find compiler -name "*.cpp" -exec tabs2spaces {} \;

Repeat this, replacing "*.cpp" with "*.h" and "*.c", and
"compiler" with "runtime"


I haven't yet set up a cron job to make sure new tabs aren't
introduced into the code, but will hope to do so tomorrow.

The command:

    cvs diff -t <filename>

may also be of use as it tends to show the diffing lines
in a clearer manner (but does not remove the fact that
there's a difference).


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@2006 3a8e244f-b0f2-452b-bcba-4c88e055c3ca