This is a checkin of code that is not used by default.  There is a new commented out #define in passlist.h called USE_EXPR_VARINIT which if uncommented will use VarInitExprs as discussed below.
Added a new pass called ProcessDefs that takes VarDefStmts and inserts
default and user initialization statements into the code.  This is
done before analysis.  Since the type of the variable may not be
known, the expr VarInitExpr of the VarSymbol is used to mark that
var->type->defaultVal should be used here when the defaultVal is
known.

After analysis, this replacement is done.  It is currently done in
RemoveNamedParameters.

I'm checking in so that John can get analysis going when he gets a
chance.  I think this is going to require a little work on my end so
I'm using the #define method.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@2025 3a8e244f-b0f2-452b-bcba-4c88e055c3ca