Fixed a bug in sequence types in which the type had astType TYPE_CLASS rather than TYPE_SEQ.
Fixed and added pass VerifyASTType to the passlist just before
analysis is run.  This is another nice sanity check.

Reworked sequence code in the internal prelude.  Most of it remains
commented out.

Made BuildClassConstructorsEtc work on all modules, not just the
common and user modules.

Fixed a bug in specializeParenOpExprs in which a variable was not set
to NULL.  This didn't matter until there was an error in which case
the INT_FATAL was incorrectly not firing.

Added a simple test for arrays of strings.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@2035 3a8e244f-b0f2-452b-bcba-4c88e055c3ca