Added overloaded = for records.
As a stopgap until VarInitExpr is working well, I fixed how
constructors are made so that an assignment to a string is done with
the _init_string function rather than assignment.  As a stopgap to
this, I added a bail-out in ProcessParameters if the function is
called _init_string.  This let me move on with the test I was writing
and should fix the regression we've been seeing.

Added a test case for record assignment that is working even though
the overloaded = is not firing.  This is because C does it for us.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@2043 3a8e244f-b0f2-452b-bcba-4c88e055c3ca