Moved the RemoveLikeTypes traversal before analysis.  Added an error message if we cannot resolve the type of a like expression before analysis.
Improved typeInfo() on MemberAccess expressions so that if it is run
before analysis it checks to see if it can resolve the type anyway.
Alternative: This can be moved to ScopeResolve symbols in this case.

Removed futures for like-1 and like-2.  Fixed good output file of
like-1.  Hopefully this doesn't break our test-case policy.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@2071 3a8e244f-b0f2-452b-bcba-4c88e055c3ca