Implemented the --memtrace and --memthreshold runtime options. "--memtrace=<filename>" prints a memory trace to the given filename.  For instance:
        "malloc" called for "49" items of size "8" for \
                 "array A":  0x804f300

The memory threshold option is always called with memory trace,
"--memtrace=<filename> --memthreshold=<n>", and squelches prints
to the log for memory sizes less than 'n'.

Also updated the help message, sorting the flags by type and
printing them under headings.  The config var table is printed
at the bottom of the help message.

Changed the print of the memory table to include both the number
and size (in bytes) of the memory requested, along with the
memory description.

Added functionality to the test harness so that a test can
specify its own CATFILES, e.g. foo.catfiles.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@2086 3a8e244f-b0f2-452b-bcba-4c88e055c3ca