Changed VarInitExpr to contain an Expr rather than a Symbol so that it can contain MemberAccesses when used in class constructors.
Reworked BuildClassConstructors to use VarInitExpr iff

      #define USE_VAR_INIT_EXPR

is uncommented.  Otherwise it uses the type->defaultVal if it is known
and nothing otherwise.

Improved support for sequences.  I believe I'm close to getting these
things working.  Create a new function createSeqType that grabs the
_seq class in the internal prelude and turns it into a new sequence
type when requested to do so by InsertAnonymousTypes.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@2114 3a8e244f-b0f2-452b-bcba-4c88e055c3ca