This is the first in a series of checkins to get operator= working. I am going to try to do it this way as integrating this and VarInitExpr was becoming too much.
The main change here is that all unknown variables are initialized to nil.

NOTE: when we get VarInitExpr in this is going to change, so if it seems
a bit messy now don't worry about it, this is transition code.

Updated 2 tests which were effected.  Added unistd.h to stdchpl.h
so as to get intptr_t.  Changed codeGen of assignment to handle
assigning numbers from nil.  Again, this is transitional code, please
bear with me.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@2120 3a8e244f-b0f2-452b-bcba-4c88e055c3ca