Made major improvements to sequences.  Rather than withing in the base _seq class to create the anonymous sequence type, the class is cloned (methods and all).
Fixed a number of bugs in update in copy.  This will help us with
class cloning eventually.

Modified the code for _seq.  This checkin gets test_seq1b and
test_seq1c working with --noanalysis.

Updated the hard-coded write function for sequences to use nil instead
of NULL because nil != NULL.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@2126 3a8e244f-b0f2-452b-bcba-4c88e055c3ca