Moved ExpandSeqExprAssignments which transforms something like
  s = (/ 1, 3, 4 /);

into something like

  s = s # 1;
  s = s # 3;
  s = s # 4;

though "#" is not yet supported all the way down in the passlist (now
after analysis).  This means analysis will need to deal with SeqExpr
(sequence literals of the form (/ ... /)) though this was to be
expected.

Added test_seq1 and test_seq1a to my noanalysis sequence testing
directory as these are now working with --noanalysis.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@2135 3a8e244f-b0f2-452b-bcba-4c88e055c3ca