Added a new pass RemoveSeqOperations that converts "#" into a method call into the sequence class.  This is necessarily done after analysis since before analysis we do not necessarily know whether the left or right hand expression is a sequence (as opposed to an element) or if they are both sequences.
Added non-analysis support for "#" used to append an element to a
sequence.  This uses the BINOP_SEQCAT that already existed but was
never used.

Fixed a bug in ApplyThisParameters in which constructors were getting
a "this" wrapped around them.  That seems strange since we don't want
to move the "this" into the constructor.  I'm a little surprised it
was working.  Must not have been happening in general.

Removed some dead code from the function that creates sequence types
from class types.

Removed old runtime support macro _SEQ_APPEND.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@2141 3a8e244f-b0f2-452b-bcba-4c88e055c3ca