Made significant progress on support for sequences.  Added more support for the # operator on two sequences or an element and a sequence, not just a sequence and an element.
Implemented a new pass to insert temps for all sequence literals that
are not in the top-level of a DefStmt (and thus the temp would be
pointless).  This pass is called InsertTemps as it will eventually be
used to insert temps for tuple literals as well.

Made two more ugly hacks to deal with the __init_function being
different from all other functions.  I'll deal with this soon.

Made a hack to retype indices in for loops after analysis if the for
loop is going over a sequence.  Change the dtInteger to whatever the
sequences elementType is.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@2165 3a8e244f-b0f2-452b-bcba-4c88e055c3ca