Reinserted redefinition checking in the symboltable.
Reworked handling of getter functions with the same name as the field.
We first name the getter "_chplget_$field" where $field is the name of
the field.  Then, later, we rename it "$field."  This means that it
can no longer be looked up in the Symboltable.  However, we don't
care.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@2173 3a8e244f-b0f2-452b-bcba-4c88e055c3ca