Fixed a bunch of compiler warnings that show up with
  -O1 -Wuninitialized -Winit-self

all of which I believe turned out to be false positives.

Removed John's

  #define CALL_RETURN_TYPE_INFO

replacing it with an internal warning if the return type is ever not
known at codegen time.  I saw this in program

  functions/deitz/test_default_param1.chpl

not

  functions/deitz/test_named_param3.chpl.

Moreover, when I ran it in gdb or with valgrind, the error
disappeared.  I'm no longer seeing the error with the above innocuous
changes leading me to believe that there is some strange
initialization error.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@2212 3a8e244f-b0f2-452b-bcba-4c88e055c3ca