Added an init function to the common module that is called at the beginning of the program.  This extends the functionality of the common module.
This checkin contains a patch of SymtabTraversal over the Common and
USER modules so that symbols are not traversed twice, care of Brad.

Set the defaultVal of arrays to zero mostly as a stopgap although this
is also where we're headed.

Removed "static" from variables that are declared in the common
module since these are "exported."

Refuturized test_seq6 and test_seq7 until sequence initialization is
figured out.

Updated flowanalysis ast output to take into account changes to init
in the common module.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@2224 3a8e244f-b0f2-452b-bcba-4c88e055c3ca