Made Expr::typeInfo calls work before analysis, though very conservatively.
As a result of this change, I updated the output to
  test/flowanalysis/newast/pareninit.good
John, can you make sure this is okay.  It almost looks like integer is
known now.

Started building replaceExpr functions to replace children pointers of
expressions.  These are currently not used but should eventually help
replace the "Expr** back" pointer to a pointer.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@2285 3a8e244f-b0f2-452b-bcba-4c88e055c3ca