Update several .future files and remove a duplicate \n from the end of warnings/errors.
In particular, change a .future to note:

The problem is that Tuple is a class,
not a TupleType.  In order to make the change
we need to be able to create methods on TupleType's
so that the tuple indexing function can be built
in internal_prelude.chpl.   Alternatively, tuple
indexing could be auto-built along with the other
setters, getters.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@2296 3a8e244f-b0f2-452b-bcba-4c88e055c3ca