Renamed Expr::parent to Expr::parentExpr to better say that this is the parent of this expression and it is an Expr.  Renamed Expr::stmt to Expr::parentStmt to better say that this is the parent of this expression and it is a Stmt.
Improved Fixup and added a new test (currently commented out) to make
sure that the parentStmt of a parentExpr of an Expr is the same as the
parentStmt of that Expr.  Also added a new test (currently commented
out) to make sure that the parentSymbol of the parentStmt of a Stmt is
the same as the parentSymbol of that Stmt.  These fail but motivate me
to fixup the __init function once and for all.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@2303 3a8e244f-b0f2-452b-bcba-4c88e055c3ca