Updated compiler's Makefiles so that a version of the compiler could be made without using the BoehmGC package.  This is done using the "make nogc" option from any of the chapel/ compiler/ or compiler/* subdirectories.
Building in this mode creates object files of the form
$(PLATFORM)/expr.nogc.o, so it should be possible to
switch between gc and no-gc builds without doing any
cleaning steps in-between.  The compiler that's
produced in this compilation mode is currently called
chpl-nogc.

Next step in this effort is to make the testing system
when run with valgrind use the chpl-nogc compiler by
default.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@2375 3a8e244f-b0f2-452b-bcba-4c88e055c3ca