Reworked TupleType to be a subclass of Structural Type.  It is now treated very similarly to records.  When a type is added to TupleType using the addType procedure, a field named _field# is added to the structural type.
TupleTypes have accessors and constructors now.  There is no longer a
->defaultVal associated with TupleTypes as initialization is now done
with a defaultConstructor much like with RecordType.

I put in code to print out the old TupleType IO for IndexTypes as a
stopgap until IndexTypes use fully-defined tuples, TupleTypes with
symbols so they can be codegened and have constructors, etc.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@2426 3a8e244f-b0f2-452b-bcba-4c88e055c3ca