"class Tuple" is now "record Tuple" sincce Tuples have value semantics. The overloaded = on Index now assigns to and from integers only.
We are going to have to build special versions of overloaded =
for Index types which have an idxType which is a Tuple (as we
have for tuples).  We are also going have to add either accessors
for the "fieldN" accessors to Index types which are of Tuples or
switch the Tuple = function to use numeric indexing functions.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@2428 3a8e244f-b0f2-452b-bcba-4c88e055c3ca