Added a new static function to TypeSymbol
  TypeSymbol*
  TypeSymbol::lookupOrDefineTupleTypeSymbol(Vec<Type*>* components)

that either finds an existing tuple with the components passed or
creates one.  This will currently only work on tuples of primitive
types because the new tuple is created in _CommonModule.

This function is used in InsertAnonymousTypes when inserting anonymous
tuple types and is intended to be used by analysis to create new tuple
types on the fly.

This function builds all the standard tuple methods, like the
constructor, when it is called.

Modified BuildClassConstructorsEtc not to build the default structural
type methods if they have already been built, since they have for
tuples now.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@2430 3a8e244f-b0f2-452b-bcba-4c88e055c3ca