Replaced
  #define CONSTRUCTOR_WITH_PARAMETERS

with compiler flag --useNewConstructor to set global variable
useNewConstructor.

Added a copy of test_class3 that uses the named parameters to the
constructor and is tested with this flag to make sure that keeps
working.

Made a superficial change to the parser that will eventually help
allow declarations of the form

  var (i, j) = tuple;

though the implementation still needs to be done.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@2449 3a8e244f-b0f2-452b-bcba-4c88e055c3ca