Corrected bugs in the coercion and default parameter wrapper code.
Added a new test_coerce test that would have caught the bug and added
a copy of test_class1 that uses the --useNewConstructor flag to keep
it from regressing again with this flag on.

Updated regressions.  John can you take a look at a couple of these.
The default wrapper function looks like it is being called on the set
accessor.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@2457 3a8e244f-b0f2-452b-bcba-4c88e055c3ca