Net Result: All DefExprs now contain a single Symbol.
Previously, multiple VarSymbols were occasionally stored in the ->sym
field of a single DefExpr.

This checkin required some changes to the indices of ForallExpr and
ForLoopStmt since they now have to be a list of DefExprs rather than a
single DefExpr with a list of symbols.

Updated printing outputs of Forall expressions which now show type
information of the index variables.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@2477 3a8e244f-b0f2-452b-bcba-4c88e055c3ca