Started integrating my index type insertion with Steve's dtInteger assumption. The idea is to make the transition smoother and wait for changes to forall loops to handle at least tuples of integers, rather that integer symbols.
My version will work when the _dtInteger_IndexType_switch is
turned on. It will break most of the tests that are working under
Steve's assumption, but it will pass some tests that are not. For any
purpose other than local playing around with it, this flag should
be off.

For now there is not much to play with -- checking in just to
avoid Steve's changes (typically to the same lines of code
I work on) mangling my code.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@2539 3a8e244f-b0f2-452b-bcba-4c88e055c3ca