Added a new field
  AssignOp* DefExpr::init

that will eventually replace Expr* VarSymbol::init in assignment form
where the rhs is VarSymbol::init and the lhs is Variable of the
VarSymbol being defined in the DefExpr.

I'm having trouble getting analysis to traverse this expression so,
John, could you take a look at it.  I added a runtime flag
--useNewInit for switching over to this.  This flag fills in this
expression, but VarSymbol::init remains as well.

The major challenge to making this work was in Fixup to get it to
maintain the parentExpr and parentStmt of DefExpr::init, even though
it will be lost for the definitions of the symbols.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@2544 3a8e244f-b0f2-452b-bcba-4c88e055c3ca